On branch Klenovsky-add-tasks  Changes to be committed: 	modified:   src/additional_tasks/task_12.py 	modified:   src/additional_tasks/task_3.py

+++ src/additional_tasks/task_3.py
@@ -2,13 +2,13 @@
 # S=1+1/2+1/3+1/4+...+1/N, где N – натуральное число.
 
 
-def some_calc(number: int) -> float:
+def some_calc(number: float) -> float:
     """
     Вычисление: S = 1 + 1 / 2 + 1 / 3 + 1 / 4 + ... + 1 / N
     :param number: Это N
     :return: Результат вычисления
     """
-    result = 0
+    result = 0.0
     for i in range(1, number + 1):
         result += 1 / i
     return result
